requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
gunicorn==20.1.0
google-cloud-aiplatform==1.7.0
google-cloud-logging==2.7.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2